refine

+++ src/p_setup.c
@@ -781,7 +781,7 @@ static void P_LoadNodes_V4 (int lump)
       for (j=0 ; j<2 ; j++)
         {
           int k;
-          no->children[j] = (unsigned int)(mn->children[j]);
+          no->children[j] = LONG(mn->children[j]);
           for (k=0 ; k<4 ; k++)
             no->bbox[j][k] = SHORT(mn->bbox[j][k])<<FRACBITS;
         }
